+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Fixed errors in API spec, ensured consistency over all endpoints and against swagger.json
 - Fixed missing notication toast when tasks is set to done or not-done
 - Added context propagation to all database operations, added server timeout to trigger context termination and added graceful shutdown
+- Added dedicated API endpoint to move issues to another project, only admin roles have permission
 
 # 1.2.1